// src/generated/api/models/cacheConfigUpdate.gen.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * ABCo API
 * OpenAPI spec version: 0.1.0
 */
import type { CacheConfigUpdateDefaults } from './cacheConfigUpdateDefaults.gen';
import type { CacheHistoryRetentionUpdate } from './cacheHistoryRetentionUpdate.gen';

export interface CacheConfigUpdate {
  defaults?: CacheConfigUpdateDefaults;
  history_retention?: CacheHistoryRetentionUpdate | null;
}